requirements.txt
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.29.1
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.0.4
WTForms==3.1.2
